src/tasks/task-4/telegram-bot/requirements.txt
# pysqlite3-binary
pandas==2.2.3
protobuf==5.28.3
chromadb==0.5.18
python-telegram-bot==21.7
google-generativeai==0.8.3
torch==2.5.1
torchaudio==2.5.1
deep-translator==1.11.4
transformers==4.46.2
gtts==2.5.4a
pyctcdecode==0.5.0